Extract depth camera-parameters prompt/dialog cluster into engine/src/depthCameraParamsPrompt.ts
Moves promptForCameraParameters, saveCameraParams, handleCameraParams,
handleCameraParamsCancelled, and handleCameraParamsError out of main.ts.
